requirements.txt
anthropic==0.97.0
rank-bm25==0.2.2
pydantic==2.13.3
python-dotenv==1.2.2
